Created and added pokemonInterfaces file
